+++ Lutra/Lutra.Infrastructure/Migrations/20260917195720_EnlargeVerspakketNaam.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using Lutra.Infrastructure.Sql;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 
@@ -11,9 +12,11 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 namespace Lutra.Infrastructure.Sql.Migrations
 {
     [DbContext(typeof(LutraDbContext))]
-    partial class LutraDbContextModelSnapshot : ModelSnapshot
+    [Migration("20260917195720_EnlargeVerspakketNaam")]
+    partial class EnlargeVerspakketNaam
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder